Version 0.3 /Splash and Login with All, ShopDetails, Firebase Storage Implements to load images

+++ app/src/main/java/com/example/rapicarmen/Shop.kt
@@ -6,16 +6,15 @@ import com.google.firebase.firestore.IgnoreExtraProperties
 @Keep
 @IgnoreExtraProperties
 data class Shop(
-
     private var nombre: String? = null,
-    private var telefono: Int? = null,
+    private var telefono: Long? = null,
     private var categoria: String? = null) {
 
     fun getNombre(): String{return this.nombre!!}
     fun setNombre(name: String){this.nombre = name}
 
-    fun getTelefono(): Int {return this.telefono!!}
-    fun setTelefono(tel: Int){this.telefono = tel}
+    fun getTelefono(): Long {return this.telefono!!}
+    fun setTelefono(tel: Long){this.telefono = tel}
 
     fun getCategoria(): String {return this.categoria!!}
     fun setCategoria(cat: String){this.categoria = cat}
